+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed:
-- Allow setting active until field in habits
+- Allow setting habit active or inactive
 
 ## [0.8.190] - 2022-11-04
 - Flutter upgrade